restructure canvas submodule
