+++ src/modules/osubot/canvas/index.js
@@ -1,0 +1,5 @@
+import drawRecent from './recent'
+import drawStat from './stat'
+import drawBest from './best'
+import avatar from './avatar'
+export default { drawRecent, drawStat, drawBest, ...avatar}
